@@ -1,4 +1,4 @@
-// +build tools
+//go:build tools
 
 // Tooling dependencies
 // https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module
